Refactored app and added routing
