+++ src/components/SuiteCard/index.js
@@ -1,0 +1,34 @@
+import React from "react";
+import './index.scss';
+import { Link } from 'react-router-dom';
+import propTypes from 'prop-types';
+
+const SuiteCard = props => (
+  <div className="col-3">
+    <div className="card m-1">
+      <Link to={`/suites/${props.id}`}>
+        <img
+          className="card-img-top"
+          src="https://via.placeholder.com/362x200"
+          alt="Suite card"
+        />
+        <div className="card-body">
+          <h5 className="card-title">{props.title}</h5>
+          <p className="card-text"> Size: {props.size} ft</p>
+          <p className="card-text"> Rooms: {props.rooms}</p>
+          <p className="card-text"> Den: {props.den}</p>
+          <p className="card-text"> Description: {props.description}</p>
+        </div>
+      </Link>
+    </div>
+  </div>
+)
+
+SuiteCard.propTypes = {
+  id: propTypes.string.isRequired,
+  size: propTypes.number.isRequired,
+  rooms: propTypes.string.isRequired,
+  den: propTypes.string.isRequired
+}
+
+export default SuiteCard;
